Replace p2panda-stream with -next
