+++ Cargo.toml
@@ -13,7 +13,6 @@ members = [
     "p2panda-store",
     "p2panda-store-next",
     "p2panda-stream",
-    "p2panda-stream-next",
     "p2panda-sync",
 ]
