+++ p2panda-stream/Cargo.toml
@@ -2,16 +2,6 @@
 name = "p2panda-stream"
 version = "0.5.1"
 edition = "2024"
-authors = [
-  "adz <x12@adz.garden>",
-  "sandreae <contact@samandreae.com>",
-  "glyph <glyph@mycelial.technology>",
-]
-description = "Stream-based methods to conveniently handle p2panda operations"
-repository = "https://github.com/p2panda/p2panda"
-license = "MIT OR Apache-2.0"
-readme = "README.md"
-keywords = ["stream", "async"]
 
 [package.metadata.docs.rs]
 all-features = true
@@ -19,22 +9,27 @@ all-features = true
 [lints]
 workspace = true
 
+[features]
+default = ["orderer"]
+orderer = [
+  "dep:p2panda-core",
+  "dep:p2panda-store-next",
+  "dep:thiserror",
+]
+
 [dependencies]
-ciborium = "0.2.2"
-futures-channel = "0.3.31"
-futures-util = { version = "0.3.31", features = ["sink"] }
-p2panda-core = { path = "../p2panda-core", version = "0.5.1", features = ["prune"] }
-p2panda-store = { path = "../p2panda-store", version = "0.5.1" }
-# @TODO: This will soon replace `p2panda-store`.
-p2panda-store-next = { path = "../p2panda-store-next", version = "0.5.1", default-features = false }
+futures-core = { version = "0.3.31", default-features = false, features = ["alloc"] }
+p2panda-core = { path = "../p2panda-core", version = "0.5.1", optional = true }
+p2panda-store-next = { path = "../p2panda-store-next", version = "0.5.1", optional = true }
 pin-project = "1.1.10"
-pin-utils = "0.1.0"
-thiserror = "2.0.12"
+thiserror = { version = "2.0.16", optional = true }
+tokio = { version = "1.47.1", default-features = false, features = ["rt", "sync", "macros"] }
 
 [dev-dependencies]
-async-stream = "0.3.6"
-p2panda-store = { path = "../p2panda-store", version = "0.5.1", features = ["sqlite", "test_utils"] }
-p2panda-store-next = { path = "../p2panda-store-next", version = "0.5.1", features = ["sqlite", "test_utils"], default-features = false }
-serde = { version = "1.0.219", features = ["derive"] }
-tokio = { version = "1.44.2", features = ["rt", "macros"] }
-tokio-stream = "0.1.17"
+futures-test = { version = "0.3.31" }
+futures-util = { version = "0.3.31", default-features = false, features = ["alloc"] }
+p2panda-core = { path = "../p2panda-core", version = "0.5.1" }
+p2panda-store-next = { path = "../p2panda-store-next", version = "0.5.1", features = ["test_utils"] }
+serde = { version = "1.0.219" }
+tokio = { version = "1.47.1", features = ["macros", "rt", "sync", "time"] }
+tokio-stream = { version = "0.1.17", default-features = false }
