--- p2panda-stream/src/macros.rs
@@ -1,71 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-/// Forward `Sink` to underlying `Stream`.
-macro_rules! delegate_sink {
-    ($field:ident, $item:ty) => {
-        fn poll_ready(
-            self: core::pin::Pin<&mut Self>,
-            cx: &mut core::task::Context<'_>,
-        ) -> core::task::Poll<Result<(), Self::Error>> {
-            self.project().$field.poll_ready(cx)
-        }
-
-        fn start_send(self: core::pin::Pin<&mut Self>, item: $item) -> Result<(), Self::Error> {
-            self.project().$field.start_send(item)
-        }
-
-        fn poll_flush(
-            self: core::pin::Pin<&mut Self>,
-            cx: &mut core::task::Context<'_>,
-        ) -> core::task::Poll<Result<(), Self::Error>> {
-            self.project().$field.poll_flush(cx)
-        }
-
-        fn poll_close(
-            self: core::pin::Pin<&mut Self>,
-            cx: &mut core::task::Context<'_>,
-        ) -> core::task::Poll<Result<(), Self::Error>> {
-            self.project().$field.poll_close(cx)
-        }
-    };
-}
-
-/// Common methods for `Stream` or `Sink` implementations to acquire pinned mutable access to
-/// projected inner fields.
-macro_rules! delegate_access_inner {
-    ($field:ident, $inner:ty, ($($ind:tt)*)) => {
-        /// Acquires a reference to the underlying sink or stream that this combinator is pulling
-        /// from.
-        pub fn get_ref(&self) -> &$inner {
-            (&self.$field) $($ind get_ref())*
-        }
-
-        /// Acquires a mutable reference to the underlying sink or stream that this combinator is
-        /// pulling from.
-        ///
-        /// Note that care must be taken to avoid tampering with the state of the sink or stream
-        /// which may otherwise confuse this combinator.
-        pub fn get_mut(&mut self) -> &mut $inner {
-            (&mut self.$field) $($ind get_mut())*
-        }
-
-        /// Acquires a pinned mutable reference to the underlying sink or stream that this
-        /// combinator is pulling from.
-        ///
-        /// Note that care must be taken to avoid tampering with the state of the sink or stream
-        /// which may otherwise confuse this combinator.
-        pub fn get_pin_mut(self: core::pin::Pin<&mut Self>) -> core::pin::Pin<&mut $inner> {
-            self.project().$field $($ind get_pin_mut())*
-        }
-
-        /// Consumes this combinator, returning the underlying sink or stream.
-        ///
-        /// Note that this may discard intermediate state of this combinator, so care should be
-        /// taken to avoid losing resources when this is called.
-        pub fn into_inner(self) -> $inner {
-            self.$field $($ind into_inner())*
-        }
-    }
-}
-
-pub(crate) use {delegate_access_inner, delegate_sink};
