--- p2panda-stream/src/operation.rs
@@ -1,285 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-//! Methods to handle p2panda operations.
-use p2panda_core::{
-    Body, Extensions, Header, Operation, OperationError, validate_backlink, validate_operation,
-};
-use p2panda_store::{LogStore, OperationStore};
-use thiserror::Error;
-
-/// Checks an incoming operation for log integrity and persists it into the store when valid.
-///
-/// This method also automatically prunes the log when a prune flag was set.
-///
-/// If the operation seems valid but we're still lacking information (as it might have arrived
-/// out-of-order) this method does not fail but indicates that we might have to retry again later.
-pub async fn ingest_operation<S, L, E>(
-    store: &mut S,
-    header: Header<E>,
-    body: Option<Body>,
-    header_bytes: Vec<u8>,
-    log_id: &L,
-    prune_flag: bool,
-) -> Result<IngestResult<E>, IngestError>
-where
-    S: OperationStore<L, E> + LogStore<L, E>,
-    E: Extensions,
-{
-    let operation = Operation {
-        hash: header.hash(),
-        header,
-        body,
-    };
-
-    if let Err(err) = validate_operation(&operation) {
-        return Err(IngestError::InvalidOperation(err));
-    }
-
-    let already_exists = store
-        .has_operation(operation.hash)
-        .await
-        .map_err(|err| IngestError::StoreError(err.to_string()))?;
-    if !already_exists {
-        // If no pruning flag is set, we expect the log to have integrity with the previously given
-        // operation.
-        if !prune_flag {
-            let latest_operation = store
-                .latest_operation(&operation.header.public_key, log_id)
-                .await
-                .map_err(|err| IngestError::StoreError(err.to_string()))?;
-
-            match latest_operation {
-                Some(latest_operation) => {
-                    if let Err(err) = validate_backlink(&latest_operation.0, &operation.header) {
-                        match err {
-                            // These errors signify that the sequence number is monotonic
-                            // incrementing and correct, however the backlink does not match.
-                            OperationError::BacklinkMismatch
-                            | OperationError::BacklinkMissing
-                            // Log can only contain operations from one author.
-                            | OperationError::TooManyAuthors => {
-                                return Err(IngestError::InvalidOperation(err))
-                            }
-                            // We observe a gap in the log and therefore can't validate the
-                            // backlink yet.
-                            OperationError::SeqNumNonIncremental(expected, given) => {
-                                // Overflows happen when the given operation is already "outdated"
-                                // and was pruned and our log grew in the meantime (thus "latest
-                                // operation" returns a higher sequence number). This is a race
-                                // condition which can occur in this async setting.
-                                //
-                                // We can safely ignore the operation here. It will _not_ be
-                                // persisted as it technically was pruned already, we will still
-                                // forward it to the application layer though. The application can
-                                // decide if it should ignore or handle the "outdated" operation.
-                                //
-                                // TODO(adz): This implementation would show undefined behaviour
-                                // when handling forks in pruned logs in presence of this race
-                                // condition. This is a rather edge case, but needs to be dealt
-                                // with.
-                                let (behind, overflow) = given.overflowing_sub(expected);
-                                if overflow {
-                                    return Ok(IngestResult::Outdated(operation))
-                                } else {
-                                    return Ok(IngestResult::Retry(operation.header, operation.body, header_bytes, behind));
-                                }
-                            }
-                            _ => unreachable!("other error cases have been handled before"),
-                        }
-                    }
-                }
-                // We're missing the whole log so far.
-                None => {
-                    if operation.header.seq_num == 0 {
-                        // We're at the beginning of the log, it's okay if there's no previous
-                        // operation.
-                    } else {
-                        // We didn't prune, there's no previous operation, something is missing.
-                        return Ok(IngestResult::Retry(
-                            operation.header.clone(),
-                            operation.body.clone(),
-                            header_bytes,
-                            operation.header.seq_num,
-                        ));
-                    }
-                }
-            }
-        }
-
-        store
-            .insert_operation(
-                operation.hash,
-                &operation.header,
-                operation.body.as_ref(),
-                &header_bytes,
-                log_id,
-            )
-            .await
-            .map_err(|err| IngestError::StoreError(err.to_string()))?;
-
-        if prune_flag {
-            store
-                .delete_operations(
-                    &operation.header.public_key,
-                    log_id,
-                    operation.header.seq_num,
-                )
-                .await
-                .map_err(|err| IngestError::StoreError(err.to_string()))?;
-        }
-    }
-
-    Ok(IngestResult::Complete(operation))
-}
-
-/// Operations can be ingested directly or need to be re-tried if they arrived out-of-order.
-#[derive(Debug)]
-pub enum IngestResult<E> {
-    /// Operation has been successfully validated and persisted.
-    Complete(Operation<E>),
-
-    /// We're missing previous operations before we can try validating the backlink of this
-    /// operation.
-    ///
-    /// The number indicates how many operations we are lacking before we can attempt validation
-    /// again.
-    Retry(Header<E>, Option<Body>, Vec<u8>, u64),
-
-    /// Operation can be considered "outdated" as a "newer" operation in the log removed this
-    /// operation ("pruning") while we processed it.
-    ///
-    /// Applications usually want to ignore these operations as the latest operation will hold all
-    /// the state we need. Additionally we were also not able to correctly check its log integrity
-    /// anymore.
-    Outdated(Operation<E>),
-}
-
-/// Errors which can occur due to invalid operations or critical storage failures.
-#[derive(Clone, Debug, Error)]
-pub enum IngestError {
-    /// Operation can not be authenticated, has broken log- or payload integrity or doesn't follow
-    /// the p2panda specification.
-    #[error("operation validation failed: {0}")]
-    InvalidOperation(OperationError),
-
-    /// Header did not contain the extensions required by the p2panda specification.
-    #[error("missing \"{0}\" extension in header")]
-    MissingHeaderExtension(String),
-
-    /// Critical storage failure occurred. This is usually a reason to panic.
-    #[error("critical storage failure: {0}")]
-    StoreError(String),
-
-    /// Some implementations might optimistically retry to ingest operations which arrived
-    /// out-of-order. This error comes up when all given attempts have been exhausted.
-    #[error("too many attempts to ingest out-of-order operation ({0} behind in log)")]
-    MaxAttemptsReached(u64),
-}
-
-#[cfg(test)]
-mod tests {
-    use p2panda_core::{Hash, Header, PrivateKey};
-    use p2panda_store::MemoryStore;
-
-    use crate::operation::{IngestResult, ingest_operation};
-    use crate::test_utils::{Extensions, StreamName};
-
-    #[tokio::test]
-    async fn retry_result() {
-        let mut store = MemoryStore::<usize, Extensions>::new();
-        let private_key = PrivateKey::new();
-        let log_id = 1;
-
-        // 1. Create a regular first operation in a log.
-        let mut header = Header {
-            public_key: private_key.public_key(),
-            version: 1,
-            signature: None,
-            payload_size: 0,
-            payload_hash: None,
-            timestamp: 0,
-            seq_num: 0,
-            backlink: None,
-            previous: vec![],
-            extensions: Extensions::default(),
-        };
-        header.sign(&private_key);
-        let header_bytes = header.to_bytes();
-
-        let result = ingest_operation(&mut store, header, None, header_bytes, &log_id, false).await;
-        assert!(matches!(result, Ok(IngestResult::Complete(_))));
-
-        // 2. Create an operation which has already advanced in the log (it has a backlink and
-        //    higher sequence number).
-        let mut header = Header {
-            public_key: private_key.public_key(),
-            version: 1,
-            signature: None,
-            payload_size: 0,
-            payload_hash: None,
-            timestamp: 0,
-            seq_num: 12, // we'll be missing 11 operations between the first and this one
-            backlink: Some(Hash::new(b"mock operation")),
-            previous: vec![],
-            extensions: Extensions::default(),
-        };
-        header.sign(&private_key);
-        let header_bytes = header.to_bytes();
-
-        let result = ingest_operation(&mut store, header, None, header_bytes, &log_id, false).await;
-        assert!(matches!(result, Ok(IngestResult::Retry(_, None, _, 11))));
-    }
-
-    #[tokio::test]
-    async fn ignore_outdated_pruned_operations() {
-        // Related issue: https://github.com/p2panda/p2panda/issues/711
-
-        let mut store = MemoryStore::<usize, Extensions>::new();
-        let private_key = PrivateKey::new();
-        let log_id = 1;
-
-        // 1. Create an advanced operation in a log which prunes all previous operations.
-        let mut header = Header {
-            public_key: private_key.public_key(),
-            version: 1,
-            signature: None,
-            payload_size: 0,
-            payload_hash: None,
-            timestamp: 0,
-            seq_num: 1,
-            backlink: Some(Hash::new(b"mock operation")),
-            previous: vec![],
-            extensions: Extensions {
-                stream_name: StreamName::new(private_key.public_key(), None),
-                prune_flag: true.into(),
-            },
-        };
-        header.sign(&private_key);
-        let header_bytes = header.to_bytes();
-
-        let result = ingest_operation(&mut store, header, None, header_bytes, &log_id, true).await;
-        assert!(matches!(result, Ok(IngestResult::Complete(_))));
-
-        // 2. Create an operation which is from an "outdated" seq from before the log was pruned.
-        let mut header = Header {
-            public_key: private_key.public_key(),
-            version: 1,
-            signature: None,
-            payload_size: 0,
-            payload_hash: None,
-            timestamp: 0,
-            seq_num: 0,
-            backlink: None,
-            previous: vec![],
-            extensions: Extensions {
-                stream_name: StreamName::new(private_key.public_key(), None),
-                prune_flag: false.into(),
-            },
-        };
-        header.sign(&private_key);
-        let header_bytes = header.to_bytes();
-
-        let result = ingest_operation(&mut store, header, None, header_bytes, &log_id, false).await;
-        assert!(matches!(result, Ok(IngestResult::Outdated(_))));
-    }
-}
