+++ p2panda-stream/src/orderer/traits.rs
