+++ p2panda-stream/src/processors/buffered.rs
