+++ p2panda-stream/src/processors/composed.rs
