+++ p2panda-stream/src/processors/pipeline.rs
