+++ p2panda-stream/src/processors/processor.rs
