+++ p2panda-stream/src/processors/tests.rs
