--- p2panda-stream/src/stream/mod.rs
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-
-mod decode;
-mod ingest;
-
-pub use decode::{Decode, DecodeExt};
-pub use ingest::{Ingest, IngestExt};
